requirements.txt
python-telegram-bot
python-dotenv
sqlite3
reportlab
Pillow
